Accept instantiated Worker in jor1kGUI constructor
This will remove the dependency on a hard-coded worker source URL,
allowing a user to pull-in the jor1k worker code using a URL
appropriate for their project structure.

This could have been achieved by accepting a URL string directly,
however, it wouldn't have been as flexible compared to directly
accepting an instantiated Worker object. For example, it is possible
to construct a Worker using a `Blob` object instead of a separate JS
source file.