Send messages to Worker only after initialization
JavaScript has a concurrency model based on an "event loop", and a
given runtime processes messages from a message queue one by one (see
https://developer.mozilla.org/en-US/docs/Web/JavaScript/EventLoop for
more information). A web worker is a distinct runtime, separate from
the "main thread", having its own message queue.

Currently, the jor1k worker is written in a way that it is initialized
completely in a single message (because of the synchronous
CommonJS-style `require()` calls supported by Browserify). This means
any messages sent from the master (including Init) are queued up for
the worker, which are processed after initialization as expected.

The project I'm working on, https://github.com/cs-education/sysbuild,
uses RequireJS for module loading. Because RequireJS uses AMD-style
modules, we use a CommonJS Loader Plugin
(https://github.com/guybedford/cjs) to wrap the modules from Jor1k
into AMD-style modules when loading. Because AMD-style requires are
asynchronous (even after bundling with the r.js optimizer, where
`setTimeouts` with zero delays are used), the jor1k worker is not
initialized completely in single message, but instead the
initialization process is split over several messages. This means that
messages sent from the master are almost always queued up before the
worker has registered listeners for them, and so the messages sent
from master are lost. This leads to the Init message being ignored,
causing the worker to not start up the VM.

The changes in this commit ensure that the master sends the Init and
other initial messages only after the worker has finished initializing
and has registered the necessary listeners. This is implemented by
waiting for the worker to send a "WorkerReady" message to the master.
I believe performance should not be impacted due to this, as there is
only one additional message being processed, and that too only during
initialization.

At this point, the worker is not only friendly to AMD loaders, but
also does not make any assumptions about any initialization and
module-loading being synchronous. The master is also most likely
friendly to AMD loaders, as we have been successfully loading it using
RequireJS + CJS plugin in our project for a while now.